requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
influxdb-client==1.35.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
python-dateutil==2.8.2
reactivex==4.0.4
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2